Moving update functions to their packages and declaring the update function in the challenge_controller.go. Also, fixed a problem with pow and create a package for status.
