+++ kctf-operator/pkg/controller/challenge/service/functions.go
@@ -4,6 +4,7 @@ package service
 
 import (
 	"context"
+	"reflect"
 
 	"github.com/go-logr/logr"
 	kctfv1alpha1 "github.com/google/kctf/pkg/apis/kctf/v1alpha1"
@@ -11,13 +12,35 @@ import (
 	netv1beta1 "k8s.io/api/networking/v1beta1"
 	"k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/types"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 )
 
-func Create(challenge *kctfv1alpha1.Challenge, client client.Client, scheme *runtime.Scheme,
+// Check if the arrays of ports are the same
+func equalPorts(found []corev1.ServicePort, wanted []corev1.ServicePort) bool {
+	if len(found) != len(wanted) {
+		return false
+	}
+
+	for i, _ := range found {
+		if found[i].Name != wanted[i].Name || found[i].Protocol != wanted[i].Protocol ||
+			found[i].Port != wanted[i].Port || found[i].TargetPort != wanted[i].TargetPort {
+			return false
+		}
+	}
+	return true
+}
+
+// Copy ports from one service to another
+func copyPorts(found *corev1.Service, wanted *corev1.Service) {
+	found.Spec.Ports = []corev1.ServicePort{}
+	found.Spec.Ports = append(found.Spec.Ports, wanted.Spec.Ports...)
+}
+
+func create(challenge *kctfv1alpha1.Challenge, client client.Client, scheme *runtime.Scheme,
 	log logr.Logger, ctx context.Context, err_ingress error) (bool, error) {
-	serv, ingress := Generate(challenge)
+	serv, ingress := generate(challenge)
 	// Create the service
 	log.Info("Creating a new Service", "Service.Namespace",
 		serv.Namespace, "Service.Name", serv.Name)
@@ -73,7 +96,7 @@ func Create(challenge *kctfv1alpha1.Challenge, client client.Client, scheme *run
 	return true, nil
 }
 
-func Delete(serviceFound *corev1.Service, ingressFound *netv1beta1.Ingress,
+func delete(serviceFound *corev1.Service, ingressFound *netv1beta1.Ingress,
 	client client.Client, scheme *runtime.Scheme, log logr.Logger,
 	ctx context.Context, err_ingress error) (bool, error) {
 	log.Info("Deleting the Service", "Service.Namespace", serviceFound.Namespace,
@@ -100,4 +123,81 @@ func Delete(serviceFound *corev1.Service, ingressFound *netv1beta1.Ingress,
 
 	// Service deleted successfully - return and requeue
 	return true, err
+}
+
+func Update(challenge *kctfv1alpha1.Challenge, client client.Client, scheme *runtime.Scheme,
+	log logr.Logger, ctx context.Context) (bool, error) {
+	// Service is created in challenge_controller and here we just ensure that everything is alright
+	// Creates the service if it doesn't exist
+	// Check existence of the service:
+	serviceFound := &corev1.Service{}
+	ingressFound := &netv1beta1.Ingress{}
+	err := client.Get(ctx, types.NamespacedName{Name: challenge.Name,
+		Namespace: challenge.Namespace}, serviceFound)
+	err_ingress := client.Get(ctx, types.NamespacedName{Name: "https",
+		Namespace: challenge.Namespace}, ingressFound)
+
+	// Just enter here if the service doesn't exist yet:
+	if errors.IsNotFound(err) && challenge.Spec.Network.Public == true {
+		// Define a new service if the challenge is public
+		return create(challenge, client, scheme, log, ctx, err_ingress)
+
+		// When service exists and public is changed to false
+	} else if err == nil && challenge.Spec.Network.Public == false {
+		return delete(serviceFound, ingressFound, client, scheme, log, ctx, err_ingress)
+	}
+
+	// Now we check if the service and the ingress are according to the CR:
+	if challenge.Spec.Network.Public {
+		serv, ingress := generate(challenge)
+		if !equalPorts(serviceFound.Spec.Ports, serv.Spec.Ports) {
+			copyPorts(serviceFound, serv)
+			err = client.Update(ctx, serviceFound)
+			if err != nil {
+				log.Error(err, "Failed to update service", "Service Name: ",
+					serviceFound.Name, " with namespace ", serviceFound.Namespace)
+				return false, err
+			}
+			log.Info("Service updated successfully", "Name: ",
+				challenge.Name, " with namespace ", challenge.Namespace)
+			return true, nil
+		}
+		// Flags if there was a change in the ingress instance
+		change_ingress := false
+
+		// TODO: check dns
+
+		// If ingress should be created:
+		if errors.IsNotFound(err_ingress) && ingress.Spec.Backend != nil {
+			// create ingress
+			change_ingress = true
+			err = client.Create(ctx, ingress)
+		}
+
+		// Cases when the ingress should be deleted or merely updated
+		if err_ingress == nil && !reflect.DeepEqual(ingressFound.Spec, ingress.Spec) {
+			change_ingress = true
+			if ingressFound.Spec.Backend != nil && ingress.Spec.Backend == nil {
+				// Deletes ingress
+				err = client.Delete(ctx, ingressFound)
+			} else {
+				// Updates ingress
+				ingressFound.Spec = ingress.Spec
+				err = client.Update(ctx, ingressFound)
+			}
+		}
+
+		if change_ingress == true {
+			if err != nil {
+				log.Error(err, "Failed to update ingress", "Ingress Name: ",
+					ingressFound.Name, " with namespace ", ingressFound.Namespace)
+				return false, err
+			}
+			log.Info("Updated ingress successfully", "Ingress Name: ",
+				ingressFound.Name, " with namespace ", ingressFound.Namespace)
+			return true, nil
+		}
+	}
+
+	return false, nil
 }
