+++ kctf-operator/pkg/controller/challenge/status/functions.go
@@ -1,0 +1,18 @@
+package status
+
+import (
+	"context"
+
+	"github.com/go-logr/logr"
+	kctfv1alpha1 "github.com/google/kctf/pkg/apis/kctf/v1alpha1"
+	"k8s.io/apimachinery/pkg/runtime"
+	"sigs.k8s.io/controller-runtime/pkg/client"
+)
+
+func Update(challenge *kctfv1alpha1.Challenge, cl client.Client, scheme *runtime.Scheme,
+	log logr.Logger, ctx context.Context) (bool, error) {
+	// TODO
+	//challenge.Status.Status = "success"
+	//err = r.client.Status().Update(context.TODO(), challenge)
+	return false, nil
+}
